RavenDB-26662 Address PR review comments
