+++ src/Raven.Server/ServerWide/ServerStore.cs
@@ -513,7 +513,7 @@ namespace Raven.Server.ServerWide
                     var term = _engine.CurrentTerm;
                     using (ClusterMaintenanceSupervisor = new ClusterMaintenanceSupervisor(this, _engine.Tag, term))
                     using (Observer = new ClusterObserver(this, ClusterMaintenanceSupervisor, _engine, term, _engine.ContextPool, ServerShutdown))
-                    using (new WriteUsageReporter(this, Observer, term, Configuration.Licensing.WriteUsageReportingInterval.AsTimeSpan, ServerShutdown))
+                    using (new WriteUsageReporter(this, Observer, term, ServerShutdown))
                     {
                         var oldNodes = new Dictionary<string, string>();
                         while (_engine.LeaderTag == NodeTag && term == _engine.CurrentTerm)
