+++ src/Raven.Server/Web/Studio/LicenseHandler.cs
@@ -5,7 +5,9 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
+using Raven.Client.Exceptions.Commercial;
 using Raven.Server.Commercial;
+using Raven.Server.Commercial.WriteUsageMetering;
 using Raven.Server.Config.Categories;
 using Raven.Server.Json;
 using Raven.Server.Routing;
@@ -297,6 +299,45 @@ namespace Raven.Server.Web.Studio
                 var response = JsonDeserializationServer.DownloadFreeLicenseResponse(json);
 
                 context.Write(writer, response.ToJson());
+            }
+        }
+
+        [RavenAction("/license/quill/usage", "POST", AuthorizationStatus.Operator)]
+        public async Task GetQuillUsage()
+        {
+            var licenseType = ServerStore.LicenseManager.LicenseStatus.Type;
+            if (licenseType != LicenseType.Quill)
+                throw new LicenseLimitException(LimitType.Quill, $"Usage data is only available under a Quill license, but the current license type is '{licenseType}'.");
+
+            var license = ServerStore.LoadLicense();
+            if (license == null)
+                throw new LicenseLimitException(LimitType.InvalidLicense, "Usage data is only available under a Quill license, but no license is currently installed.");
+
+            using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
+            {
+                BlittableJsonReaderObject requestBody = await context.ReadForMemoryAsync(RequestBodyStream(), "usage-query");
+
+                var modifications = new DynamicJsonValue(requestBody);
+                requestBody.Modifications = modifications;
+                modifications["License"] = license.ToJson();
+                modifications["CertificateThumbprint"] = GetCurrentCertificate()?.Thumbprint;
+
+                using (var token = CreateHttpRequestBoundOperationToken())
+                using (var content = new StringContent(context.ReadObject(requestBody, "usage-query").ToString(), Encoding.UTF8, "application/json"))
+                using (var response = await ApiHttpClient.PostAsync(WriteUsageMeteringConstants.UsageQueryEndpointPath, content, HttpCompletionOption.ResponseHeadersRead, shouldRetry: false, token: token.Token).ConfigureAwait(false))
+                {
+                    HttpContext.Response.StatusCode = (int)response.StatusCode;
+
+                    var contentType = response.Content.Headers.ContentType?.ToString();
+                    if (string.IsNullOrEmpty(contentType) == false)
+                        HttpContext.Response.ContentType = contentType;
+
+                    await using (var responseStream = ResponseBodyStream())
+                    await using (var apiStream = await response.Content.ReadAsStreamAsync(token.Token))
+                    {
+                        await apiStream.CopyToAsync(responseStream, token.Token);
+                    }
+                }
             }
         }
