+++ src/Raven.Studio/typescript/test/stubs/LicenseStubs.ts
@@ -243,7 +243,6 @@ export class LicenseStubs {
             SkipLeasingErrorsLogging: false,
             LicensePath: null,
             ThrowOnInvalidOrMissingLicense: false,
-            WriteUsageReportingInterval: { AsTimeSpan: "00:15:00" },
         };
     }
